% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_leaves.R
\name{shiny_leaves}
\alias{shiny_leaves}
\title{Shiny leaves}
\usage{
shiny_leaves()
}
\description{
An app to explain how leaf area is calculated from a scan.
It can also be used to work interactively with problematic leaves to better
understand how to process them.
}
\examples{
# Run app
if (interactive()) {
 shiny_leaves()
}
}
